# apps/api/openapi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/cite/doctor/{id}": {
            "get": {
                "operationId": "CiteController_findByDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseCiteDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "cite"
                ]
            }
        },
...
        "/doctor/user/{id}": {
            "get": {
                "operationId": "DoctorController_findByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseDoctorDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            }
        },
        "/doctor/speciality/{id}": {
            "get": {
                "operationId": "DoctorController_findBySpecialityId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseDoctorDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            }
        },
        "/doctor/{id}": {
            "get": {
                "operationId": "DoctorController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDoctorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            },
            "patch": {
                "operationId": "DoctorController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoctorDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDoctorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            },
            "delete": {
                "operationId": "DoctorController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDoctorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            }
        },
        "/doctor": {
            "get": {
                "operationId": "DoctorController_findAll",
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseDoctorDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            },
            "post": {
                "operationId": "DoctorController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDoctorDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDoctorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            }
        },
        "/speciality/{id}": {
            "get": {
                "operationId": "SpecialityController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSpecialityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "speciality"
                ]
            },
            "patch": {
                "operationId": "SpecialityController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSpecialityDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSpecialityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "speciality"
                ]
            },
            "delete": {
                "operationId": "SpecialityController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSpecialityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "speciality"
                ]
            }
        },
...
        "/availability/doctor/{id}": {
            "get": {
                "operationId": "AvailabilityController_findByDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAvailabilityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "availability"
                ]
            }
        },
...
        "/record/doctor/{id}": {
            "get": {
                "operationId": "RecordController_findByDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseRecordDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "record"
                ]
            }
        },
...
        "/report/{id}": {
            "get": {
                "operationId": "ReportController_findByDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "reports"
                ]
            }
        },
...
    },
    "info": {
        "title": "MEDIGO API",
        "description": "API Endopoints for MEDIGO",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "ResponseUserPatientDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "email"
                ]
            },
...
            "ResponseSpecialityDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description"
                ]
            },
            "ResponseDoctorDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "speciality": {
                        "$ref": "#/components/schemas/ResponseSpecialityDto"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ResponseUserPatientDto"
                    }
                },
                "required": [
                    "id",
                    "phone",
                    "speciality",
                    "user"
                ]
            },
...
            "ResponsePatientDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "address": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ResponseUserPatientDto"
                    }
                },
                "required": [
                    "id",
                    "address",
                    "phone",
                    "user"
                ]
            },
            "ResponseCiteDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "date": {
                        "type": "string"
                    },
                    "time": {
                        "type": "string"
                    },
                    "patientConfirm": {
                        "type": "boolean"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/ResponseDoctorDto"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/ResponsePatientDto"
                    }
                },
                "required": [
                    "id",
                    "date",
                    "time",
                    "patientConfirm",
                    "doctor",
                    "patient"
                ]
            },
...
            "CreateDoctorDto": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    },
                    "specialityId": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "number"
                    }
                },
                "required": [
                    "phone",
                    "specialityId",
                    "userId"
                ]
            },
            "UpdateDoctorDto": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    },
                    "specialityId": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "number"
                    }
                }
            },
...
            "UpdateSpecialityDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "ResponseAvailabilityDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "start": {
                        "type": "string"
                    },
                    "end": {
                        "type": "string"
                    },
                    "day": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/ResponseDoctorDto"
                    }
                },
                "required": [
                    "id",
                    "start",
                    "end",
                    "day",
                    "available",
                    "doctor"
                ]
            },
...
            "ResponseRecordDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "description": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/ResponseDoctorDto"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/ResponsePatientDto"
                    }
                },
                "required": [
                    "id",
                    "description",
                    "date",
                    "doctor",
                    "patient"
                ]
            },
...
        }
    }
}